Refactored configs
