--- crates/solvers/src/infra/config/naive/mod.rs
@@ -1,7 +1,0 @@
-use crate::domain;
-
-pub mod file;
-
-pub struct Config {
-    pub risk: domain::Risk,
-}
